third commit - Actualización de métodos de BookController

+++ src/main/java/com/practice/obspring4restdatajpa/controllers/BootController.java
@@ -2,14 +2,19 @@ package com.practice.obspring4restdatajpa.controllers;
 
 import com.practice.obspring4restdatajpa.entities.Book;
 import com.practice.obspring4restdatajpa.repositories.BookRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class BootController {
+
+    private final Logger log = LoggerFactory.getLogger(BootController.class);
 
     private BookRepository bookRepository;
 
@@ -37,14 +42,21 @@ public class BootController {
      * @return
      */
 
-    //Buscar un libro segun su id
+    //Search a workbook by id
     @GetMapping("/api/books/{id}")
     public ResponseEntity<Book> findById(@PathVariable Long id){
 
         //Functional Programming
         return bookRepository.findById(id).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
-
 //        return bookRepository.findById(id).orElse(null);
+
+        //Ooption 2
+//        Optional<Book> bookOpt = bookRepository.findById(id);
+//        if(bookOpt.isPresent()){
+//            return ResponseEntity.ok(bookOpt.get());
+//        }else{
+//            return ResponseEntity.notFound().build();
+//        }
     }
 
 
@@ -55,22 +67,39 @@ public class BootController {
 
     //Create a new book
     @PostMapping("/api/books")
-    public Book create(@RequestBody Book book, @RequestHeader HttpHeaders headers){
-        System.out.println(headers.get("User-Agent"));
-        return bookRepository.save(book);
+    public ResponseEntity<Book> create(@RequestBody Book book, @RequestHeader HttpHeaders headers){
+        System.out.println(headers.get("User-Agent"));//This line is optional
+
+        if(book.getId() != null){
+            log.warn("Trying to create a book with an id");
+            return ResponseEntity.badRequest().build();
+        }
+        Book savedBook = bookRepository.save(book);
+        return ResponseEntity.ok(savedBook);
     }
 
 
     /**
+     * http://localhost:8080/api/books
      * http://localhost:8080/api/books/{1}
      * @return
      */
 
     //Update a book
-    @PutMapping("/api/books/{id}")
-    public Book update(@PathVariable Long id, @RequestBody Book book){
-        book.setId(id);
-        return bookRepository.save(book);
+    @PutMapping("/api/books")
+    public ResponseEntity<Book> update(@RequestBody Book book){
+//        @PathVariable Long id = /api/books/{id}
+        if (book.getId() == null){
+            log.warn("Trying to update a book without an id");
+            return ResponseEntity.badRequest().build();
+        }
+        if (!bookRepository.existsById((book.getId()))){
+            log.warn("Trying to update a non existent book");
+            return ResponseEntity.notFound().build();
+        }
+//        book.setId(id);
+        Book updateBook = bookRepository.save(book);
+        return ResponseEntity.ok(updateBook);
     }
 
 
@@ -81,7 +110,19 @@ public class BootController {
 
     //Delete a book
     @DeleteMapping("/api/books/{id}")
-    public void delete(@PathVariable Long id){
+    public ResponseEntity<Book> delete(@PathVariable Long id){
+        if (!bookRepository.existsById((id))){
+            log.warn("Trying to delete a non existent book");
+            return ResponseEntity.notFound().build();
+        }
         bookRepository.deleteById(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    @DeleteMapping("/api/books")
+    public ResponseEntity<Book> deleteAll(){
+        log.info("REST request for delete all books");
+        bookRepository.deleteAll();
+        return ResponseEntity.noContent().build();
     }
 }
